requirements_azure.txt
# Azure Functions deployment for BOL automation
PyPDF2==3.0.1
pyodbc==5.0.1
azure-functions==1.18.0
azure-functions-worker==1.0.0
requests==2.31.0
